Make underlying type of Vec3Widget a dvec3

+++ src/application/qt/widgets/vec3_widget.hpp
@@ -16,26 +16,26 @@ public:
     void setLabel(const QString& label);
     QString label() const;
 
-    void setValue(glm::vec3 value, bool force = false);
-    glm::vec3 value();
+    void setValue(glm::dvec3 value, bool force = false);
+    glm::dvec3 value();
 
-    void setRangeMin(glm::vec3 min);
-    void setRangeMax(glm::vec3 max);
-    glm::vec3 min() const;
-    glm::vec3 max() const;
-    void setStep(glm::vec3 step);
-    glm::vec3 step() const;
-    void setSingleStep(glm::vec3 step);
-    glm::vec3 singleStep() const;
-    void setPageStep(glm::vec3 step);
-    glm::vec3 pageStep() const;
+    void setRangeMin(glm::dvec3 min);
+    void setRangeMax(glm::dvec3 max);
+    glm::dvec3 min() const;
+    glm::dvec3 max() const;
+    void setStep(glm::dvec3 step);
+    glm::dvec3 step() const;
+    void setSingleStep(glm::dvec3 step);
+    glm::dvec3 singleStep() const;
+    void setPageStep(glm::dvec3 step);
+    glm::dvec3 pageStep() const;
     void setDecimals(int decimals);
     int decimals() const;
     void setAccelerated(bool accelerated);
     bool isAccelerated() const;
 
 signals:
-    void valueChanged(glm::vec3 color);
+    void valueChanged(glm::dvec3 color);
 
 private:
     struct Vec3WidgetPrivate;
